update blank app
