+++ applications/blank/custom_pde.h
@@ -1,50 +1,12 @@
 // SPDX-FileCopyrightText: © 2025 PRISMS Center at the University of Michigan
 // SPDX-License-Identifier: GNU Lesser General Public Version 2.1
 
-#pragma once
-
-#include <prismspf/core/pde_operator.h>
-#include <prismspf/core/phase_field_tools.h>
-#include <prismspf/core/variable_attribute_loader.h>
-#include <prismspf/core/variable_attributes.h>
-
-#include <prismspf/user_inputs/user_input_parameters.h>
-
-#include <prismspf/utilities/utilities.h>
-
-#include <prismspf/config.h>
+#include <prismspf/core/pde_operator_base.h>
 
 PRISMS_PF_BEGIN_NAMESPACE
 
-/**
- * @brief This is a derived class of `VariableAttributeLoader` where the user implements
- * their variable attributes and field declarations.
- */
-class CustomAttributeLoader : public VariableAttributeLoader
-{
-public:
-  /**
-   * @brief Destructor.
-   */
-  ~CustomAttributeLoader() override = default;
-
-  /**
-   * @brief User-implemented method where the variable attributes are set for all fields.
-   */
-  void
-  load_variable_attributes() override;
-};
-
-/**
- * @brief This is a derived class of `MatrixFreeOperator` where the user implements their
- * PDEs.
- *
- * @tparam dim The number of dimensions in the problem.
- * @tparam degree The polynomial degree of the shape functions.
- * @tparam number Datatype to use. Either double or float.
- */
 template <unsigned int dim, unsigned int degree, typename number>
-class CustomPDE : public PDEOperator<dim, degree, number>
+class CustomPDE : public PDEOperatorBase<dim, degree, number>
 {
 public:
   using ScalarValue = dealii::VectorizedArray<number>;
@@ -53,81 +15,39 @@ public:
   using VectorValue = dealii::Tensor<1, dim, dealii::VectorizedArray<number>>;
   using VectorGrad  = dealii::Tensor<2, dim, dealii::VectorizedArray<number>>;
   using VectorHess  = dealii::Tensor<3, dim, dealii::VectorizedArray<number>>;
-  using PDEOperator<dim, degree, number>::get_user_inputs;
-  using PDEOperator<dim, degree, number>::get_pf_tools;
-  using PDEOperator<dim, degree, number>::get_timestep;
+  using PDEOperatorBase<dim, degree, number>::get_user_inputs;
+  using PDEOperatorBase<dim, degree, number>::get_pf_tools;
 
-  /**
-   * @brief Constructor.
-   */
   explicit CustomPDE(const UserInputParameters<dim> &_user_inputs,
                      PhaseFieldTools<dim>           &_pf_tools)
-    : PDEOperator<dim, degree, number>(_user_inputs, _pf_tools)
-  {}
+    : PDEOperatorBase<dim, degree, number>(_user_inputs, _pf_tools) {};
 
 private:
-  /**
-   * @brief User-implemented class for the initial conditions.
-   */
   void
-  set_initial_condition(const unsigned int       &index,
-                        const unsigned int       &component,
-                        const dealii::Point<dim> &point,
-                        number                   &scalar_value,
-                        number                   &vector_component_value) const override;
+  set_initial_condition([[maybe_unused]] const unsigned int       &index,
+                        [[maybe_unused]] const unsigned int       &component,
+                        [[maybe_unused]] const dealii::Point<dim> &point,
+                        [[maybe_unused]] number                   &scalar_value,
+                        [[maybe_unused]] number &vector_component_value) const override {
+  };
 
-  /**
-   * @brief User-implemented function for Dirichlet boundary conditions.
-   */
   void
-  set_dirichlet(const unsigned int       &index,
-                const unsigned int       &boundary_id,
-                const unsigned int       &component,
-                const dealii::Point<dim> &point,
-                number                   &scalar_value,
-                number                   &vector_component_value) const override;
+  set_dirichlet([[maybe_unused]] const unsigned int       &index,
+                [[maybe_unused]] const unsigned int       &boundary_id,
+                [[maybe_unused]] const unsigned int       &component,
+                [[maybe_unused]] const dealii::Point<dim> &point,
+                [[maybe_unused]] number                   &scalar_value,
+                [[maybe_unused]] number &vector_component_value) const override {};
 
-  /**
-   * @brief User-implemented class for the RHS of explicit equations.
-   */
   void
-  compute_explicit_rhs(
-    VariableContainer<dim, degree, number>                    &variable_list,
-    const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-    const dealii::VectorizedArray<number>                     &element_volume,
-    Types::Index solve_block) const override;
+  compute_rhs([[maybe_unused]] FieldContainer<dim, degree, number> &variable_list,
+              [[maybe_unused]] const SimulationTimer               &sim_timer,
+              [[maybe_unused]] unsigned int solve_group_id) const override {};
 
-  /**
-   * @brief User-implemented class for the RHS of nonexplicit equations.
-   */
   void
-  compute_nonexplicit_rhs(
-    VariableContainer<dim, degree, number>                    &variable_list,
-    const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-    const dealii::VectorizedArray<number>                     &element_volume,
-    Types::Index                                               solve_block,
-    Types::Index index = Numbers::invalid_index) const override;
-
-  /**
-   * @brief User-implemented class for the LHS of nonexplicit equations.
-   */
-  void
-  compute_nonexplicit_lhs(
-    VariableContainer<dim, degree, number>                    &variable_list,
-    const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-    const dealii::VectorizedArray<number>                     &element_volume,
-    Types::Index                                               solve_block,
-    Types::Index index = Numbers::invalid_index) const override;
-
-  /**
-   * @brief User-implemented class for the RHS of postprocessed explicit equations.
-   */
-  void
-  compute_postprocess_explicit_rhs(
-    VariableContainer<dim, degree, number>                    &variable_list,
-    const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-    const dealii::VectorizedArray<number>                     &element_volume,
-    Types::Index solve_block) const override;
+  compute_lhs([[maybe_unused]] FieldContainer<dim, degree, number> &variable_list,
+              [[maybe_unused]] const SimulationTimer               &sim_timer,
+              [[maybe_unused]] unsigned int solver_id) const override {};
 };
 
 PRISMS_PF_END_NAMESPACE
