--- applications/blank/equations.cc
@@ -1,58 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 PRISMS Center at the University of Michigan
-// SPDX-License-Identifier: GNU Lesser General Public Version 2.1
-
-#include "custom_pde.h"
-
-#include <prismspf/core/type_enums.h>
-#include <prismspf/core/variable_attribute_loader.h>
-#include <prismspf/core/variable_container.h>
-
-#include <prismspf/config.h>
-
-PRISMS_PF_BEGIN_NAMESPACE
-
-void
-CustomAttributeLoader::load_variable_attributes()
-{}
-
-template <unsigned int dim, unsigned int degree, typename number>
-void
-CustomPDE<dim, degree, number>::compute_explicit_rhs(
-  [[maybe_unused]] VariableContainer<dim, degree, number> &variable_list,
-  [[maybe_unused]] const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-  [[maybe_unused]] const dealii::VectorizedArray<number> &element_volume,
-  [[maybe_unused]] Types::Index                           solve_block) const
-{}
-
-template <unsigned int dim, unsigned int degree, typename number>
-void
-CustomPDE<dim, degree, number>::compute_nonexplicit_rhs(
-  [[maybe_unused]] VariableContainer<dim, degree, number> &variable_list,
-  [[maybe_unused]] const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-  [[maybe_unused]] const dealii::VectorizedArray<number> &element_volume,
-  [[maybe_unused]] Types::Index                           solve_block,
-  [[maybe_unused]] Types::Index                           index) const
-{}
-
-template <unsigned int dim, unsigned int degree, typename number>
-void
-CustomPDE<dim, degree, number>::compute_nonexplicit_lhs(
-  [[maybe_unused]] VariableContainer<dim, degree, number> &variable_list,
-  [[maybe_unused]] const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-  [[maybe_unused]] const dealii::VectorizedArray<number> &element_volume,
-  [[maybe_unused]] Types::Index                           solve_block,
-  [[maybe_unused]] Types::Index                           index) const
-{}
-
-template <unsigned int dim, unsigned int degree, typename number>
-void
-CustomPDE<dim, degree, number>::compute_postprocess_explicit_rhs(
-  [[maybe_unused]] VariableContainer<dim, degree, number> &variable_list,
-  [[maybe_unused]] const dealii::Point<dim, dealii::VectorizedArray<number>> &q_point_loc,
-  [[maybe_unused]] const dealii::VectorizedArray<number> &element_volume,
-  [[maybe_unused]] Types::Index                           solve_block) const
-{}
-
-#include "custom_pde.inst"
-
-PRISMS_PF_END_NAMESPACE
