+++ applications/blank/main.cc
@@ -1,135 +1,43 @@
 // SPDX-FileCopyrightText: © 2025 PRISMS Center at the University of Michigan
 // SPDX-License-Identifier: GNU Lesser General Public Version 2.1
 
-#include <deal.II/base/parameter_handler.h>
+#include "custom_pde.h"
 
-#include <prismspf/core/dependencies.h>
-#include <prismspf/core/field_attributes.h>
 #include <prismspf/core/parse_cmd_options.h>
-#include <prismspf/core/phase_field_tools.h>
 #include <prismspf/core/problem.h>
-#include <prismspf/core/simulation_timer.h>
-#include <prismspf/core/solve_block.h>
-#include <prismspf/core/type_enums.h>
-
-#include <prismspf/user_inputs/input_file_reader.h>
-#include <prismspf/user_inputs/user_input_parameters.h>
-
-#include <prismspf/config.h>
 
 using namespace prisms;
-
-template <unsigned int dim, unsigned int degree, typename number>
-class CustomPDE : public PDEOperatorBase<dim, degree, number>
-{
-public:
-  using ScalarValue = dealii::VectorizedArray<number>;
-  using ScalarGrad  = dealii::Tensor<1, dim, dealii::VectorizedArray<number>>;
-  using ScalarHess  = dealii::Tensor<2, dim, dealii::VectorizedArray<number>>;
-  using VectorValue = dealii::Tensor<1, dim, dealii::VectorizedArray<number>>;
-  using VectorGrad  = dealii::Tensor<2, dim, dealii::VectorizedArray<number>>;
-  using VectorHess  = dealii::Tensor<3, dim, dealii::VectorizedArray<number>>;
-  using PDEOperatorBase<dim, degree, number>::get_user_inputs;
-  using PDEOperatorBase<dim, degree, number>::get_pf_tools;
-
-  /**
-   * @brief Constructor.
-   */
-  explicit CustomPDE(const UserInputParameters<dim> &_user_inputs,
-                     PhaseFieldTools<dim>           &_pf_tools)
-    : PDEOperatorBase<dim, degree, number>(_user_inputs, _pf_tools)
-  {}
-
-private:
-  void
-  set_initial_condition([[maybe_unused]] const unsigned int       &index,
-                        [[maybe_unused]] const unsigned int       &component,
-                        [[maybe_unused]] const dealii::Point<dim> &point,
-                        [[maybe_unused]] number                   &scalar_value,
-                        [[maybe_unused]] number &vector_component_value) const override
-  {}
-
-  void
-  compute_rhs([[maybe_unused]] FieldContainer<dim, degree, number> &variable_list,
-              [[maybe_unused]] const SimulationTimer               &sim_timer,
-              [[maybe_unused]] unsigned int solve_block_id) const override
-  {
-    if (solve_block_id == 1) // lhs
-      {
-      }
-  }
-
-  void
-  compute_lhs(FieldContainer<dim, degree, number> &variable_list,
-              const SimulationTimer               &sim_timer,
-              unsigned int                         solve_block_id) const override
-  {
-    if (solve_block_id == 1) // lhs
-      {
-      }
-  }
-};
 
 int
 main(int argc, char *argv[])
 {
-  try
-    {
-      // Initialize MPI
-      dealii::Utilities::MPI::MPI_InitFinalize
-        mpi_init(argc, argv, dealii::numbers::invalid_unsigned_int);
+  // Initialize MPI
+  dealii::Utilities::MPI::MPI_InitFinalize
+    mpi_init(argc, argv, dealii::numbers::invalid_unsigned_int);
 
-      // Restrict deal.II console printing
-      dealii::deallog.depth_console(0);
+  // Restrict deal.II console printing
+  dealii::deallog.depth_console(0);
 
-      // Parse the command line options (if there are any) to get the name of the input
-      // file
-      ParseCMDOptions cli_options(argc, argv);
-      std::string     parameters_filename = cli_options.get_parameters_filename();
+  // Parse the command line options (if there are any) to get the name of the input
+  // file
+  ParseCMDOptions cli_options(argc, argv);
 
-      constexpr unsigned int dim    = 2;
-      constexpr unsigned int degree = 2;
+  constexpr unsigned int dim    = 1;
+  constexpr unsigned int degree = 1;
 
-      std::vector<FieldAttributes> fields = {FieldAttributes("n")};
-      std::vector<SolveBlock>      solve_blocks;
-      SolveBlock                   exp_block;
-      SolveBlock                   pp_block;
-      solve_blocks.push_back(exp_block);
-      solve_blocks.push_back(pp_block);
+  std::vector<FieldAttributes> field_attributes = {};
 
-      UserInputParameters<dim>       user_inputs(parameters_filename);
-      PhaseFieldTools<dim>           pf_tools;
-      CustomPDE<dim, degree, double> pde_operator(user_inputs, pf_tools);
-      Problem<dim, degree, double>   problem(fields,
-                                           solve_blocks,
-                                           user_inputs,
-                                           pf_tools,
-                                           pde_operator);
-      problem.solve();
-    }
+  std::vector<SolveBlock> solve_groups({});
 
-  catch (std::exception &exc)
-    {
-      std::cerr << '\n'
-                << '\n'
-                << "----------------------------------------------------" << '\n';
-      std::cerr << "Exception on: " << '\n'
-                << exc.what() << '\n'
-                << "Aborting!" << '\n'
-                << "----------------------------------------------------" << '\n';
-      return 1;
-    }
-
-  catch (...)
-    {
-      std::cerr << '\n'
-                << '\n'
-                << "----------------------------------------------------" << '\n';
-      std::cerr << "Unknown exception!" << '\n'
-                << "Aborting!" << '\n'
-                << "----------------------------------------------------" << '\n';
-      return 1;
-    }
+  UserInputParameters<dim>       user_inputs(cli_options.get_parameters_filename());
+  PhaseFieldTools<dim>           pf_tools;
+  CustomPDE<dim, degree, double> pde_operator(user_inputs, pf_tools);
+  Problem<dim, degree, double>   problem(field_attributes,
+                                       solve_groups,
+                                       user_inputs,
+                                       pf_tools,
+                                       pde_operator);
+  problem.solve();
 
   return 0;
 }
