+++ applications/blank/parameters.prm
@@ -1,33 +1,19 @@
 subsection Rectangular mesh
-  set x size = 64.0
-  set y size = 64.0
-  set z size = 64.0
+  set x size = 1.0
+  set y size = 1.0
+  set z size = 1.0
   set x subdivisions = 1
   set y subdivisions = 1
   set z subdivisions = 1
-  set periodic x = false
-  set periodic y = false
-  set periodic z = false
 end
 
-set global refinement = 7
-set max refinement = 7
-set min refinement = 2
-set mesh adaptivity = false
-set remeshing period = 1000
+set global refinement = 1
 
-subsection refinement criterion: 0
-  set variables         =
-  set type              =
-  set value lower bound =
-  set value upper bound =
-end
-
-set time step = 0.0001
-set number steps = 250000
+set time step = 1.0
+set number steps = 1
 
 subsection output
   set condition = EQUAL_SPACING
-  set number = 50
-  set list =
+  set number = 1
 end
+
